Created eventsUtils
