+++ src/library/eventsUtils.js
@@ -1,0 +1,11 @@
+function missingParameters(req) {
+    const { name, location,
+            description, host,
+            status, date } = req.body;
+    if (!name || !location || !description ||
+        !host || !status || !date )
+        return true;
+    return false;
+}
+
+module.exports = { missingParameters }
